The_game.py
This python file is not finished yet. Please see the 'The_game_V2(finished).py'.
That python file is finished and it reachs the goal that I wanted. And I am trying to fix this file and let it run correctly.
If u have any issue about this file ,please contact me by sending e-mail. Thanks
PS:The goal of the two python files are the same.